pyproject.toml
[tool.poetry]
name = "datahub"
version = "0.1.0"
description = ""
authors = ["keyhong <“wnhong96@gmail.com”>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
acryl-datahub = "^0.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
